Bugfix: CostItems delete if Child Sub deleted

+++ app/grails-app/controllers/com/k_int/kbplus/SubscriptionDetailsController.groovy
@@ -926,11 +926,16 @@ class SubscriptionDetailsController extends AbstractDebugController {
             def derived_subs = Subscription.findByInstanceOfAndStatusNot(delSubscription, deletedStatus)
 
             if (!derived_subs) {
-                if (delSubscription.getConsortia() && delSubscription.getConsortia() != delInstitution) {
-                    OrgRole.executeUpdate("delete from OrgRole where sub = ? and org = ?", [delSubscription, delInstitution])
+
+                if(!CostItem.findAllBySub(delSubscription)) {
+                    if (delSubscription.getConsortia() && delSubscription.getConsortia() != delInstitution) {
+                        OrgRole.executeUpdate("delete from OrgRole where sub = ? and org = ?", [delSubscription, delInstitution])
+                    }
+                    delSubscription.status = deletedStatus
+                    delSubscription.save(flush: true)
+                }else {
+                    flash.error = message(code: 'myinst.actionDeleteChildSubscription.error', default: 'Unable to delete - The selected license has attached cost items')
                 }
-                delSubscription.status = deletedStatus
-                delSubscription.save(flush: true)
 
             } else {
                 flash.error = message(code: 'myinst.actionCurrentSubscriptions.error', default: 'Unable to delete - The selected license has attached subscriptions')
